fix: without adding the BM package to CM profile
